refactor(auth): move authd binary installation logic out of of safe-api into safe-cli
