+++ safe-api/Cargo.toml
@@ -29,7 +29,6 @@ safe_app = { git = "https://github.com/maidsafe/safe_client_libs.git", branch = 
 safe_authenticator = { git = "https://github.com/maidsafe/safe_client_libs.git", branch = "master" }
 safe_core = { git = "https://github.com/maidsafe/safe_client_libs.git", branch = "master" }
 safe-nd = "~0.6.1"
-self_update = { git = "https://github.com/bochaco/self_update", branch = "feat-s3-backend" }
 serde = "1.0.91"
 serde_json = "1.0.41"
 slog = "2.1"
