+++ safe-authd/Cargo.toml
@@ -22,7 +22,7 @@ log = "0.4.6"
 quinn = "~0.3.0"
 rcgen = "0.2"
 safe-api = { path = "../safe-api" }
-self_update = { git = "https://github.com/bochaco/self_update", branch = "feat-s3-backend" }
+self_update = { git = "https://github.com/bochaco/self_update", branch = "master" }
 serde = "1.0.91"
 serde_json = "1.0.41"
 slog = "2.1"
