+++ safe-cli/Cargo.toml
@@ -25,7 +25,7 @@ relative-path = "0.4.0"
 reqwest = "0.9.22"
 rpassword = "4.0.1"
 safe-api = { path = "../safe-api" }
-self_update = { git = "https://github.com/bochaco/self_update", branch = "feat-s3-backend" }
+self_update = { git = "https://github.com/bochaco/self_update", branch = "master" }
 serde = "1.0.91"
 serde_json = "1.0.39"
 serde_yaml = "0.8.11"
